backend/agent/pyproject.toml
[project]
name = "backend-cdk"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=1.1.1",
    "strands-agents>=1.20.0",
    "strands-agents-tools>=0.2.18",
    "youtube-transcript-api>=1.2.3",
    "python-dotenv>=1.2.1",
    "aws-opentelemetry-distro>=0.15.0",
]
